refactor(evm): gate block context behind monad
Move BlockContext into monad.rs and gate its backend helpers, replay inputs, and tests behind the monad feature. Preserve ordinary replay and synthetic transaction context in builds without Monad.
