+++ .changelog/monad-block-context-gating.md
@@ -1,0 +1,5 @@
+---
+foundry-evm-core: patch
+---
+
+Moved block-context APIs into the Monad module and gated them behind the `monad` feature.
